pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "maya-mcp"
version = "0.3.0"
description = "MCP server for controlling Autodesk Maya via commandPort"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10,!=3.10.0"
authors = [
    { name = "Maya MCP Contributors" }
]
keywords = ["maya", "mcp", "autodesk", "3d", "animation", "vfx"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: 3D Modeling",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "fastmcp>=3.2.4,<4",
    "typing-extensions>=4.12.0",
]

[project.optional-dependencies]
# Minimal lint dependencies (for CI)
lint = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]
# Minimal test dependencies (for CI)
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
# Documentation dependencies
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.25.0",
]
# Full dev dependencies (local development)
dev = [
    "maya-mcp[lint,test,docs]",
]

[project.urls]
Homepage = "https://github.com/GimbalGoats/GG_MayaMCP"
Documentation = "https://github.com/GimbalGoats/GG_MayaMCP/blob/main/README.md"
Repository = "https://github.com/GimbalGoats/GG_MayaMCP"
Issues = "https://github.com/GimbalGoats/GG_MayaMCP/issues"

[project.scripts]
maya-mcp = "maya_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/maya_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/docs",
]

# Ruff configuration
[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG001",  # Unused function argument (pytest fixtures are often dependencies)
    "ARG002",  # Unused method argument (pytest fixtures are often dependencies)
]

[tool.ruff.lint.isort]
known-first-party = ["maya_mcp"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

# Mypy configuration
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
show_error_codes = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
# Maya panel uses dynamic Qt imports (PySide2/PySide6) and Maya-specific classes
# that cannot be type-checked outside Maya
module = "maya_mcp.maya_panel.*"
ignore_errors = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-v",
    "--tb=short",
    "-ra",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "integration: marks tests as integration tests requiring Maya (deselect with '-m \"not integration\"')",
]
